EMBARGOED: perf intel-pt: Configure Trigger Tracing
Parse and validate aux-output config terms and update aux_output_cfg
accordingly.

Trigger Actions are specified by config term aux-output. Acceptable
strings (depending on capabilities) are:

on-overflow:trigger		On PMC overflow, TRIG with IP attribution
on-overflow:trigger-pause    	On PMC overflow, pause, emit TRIG with ip
on-overflow:trigger-resume	On PMC overflow, resume, emit TRIG with ip
on-overflow:trigger-noip	On PMC overflow, emit TRIG packet
on-overflow:trigger-noip-pause	On PMC overflow, pause and emit TRIG
on-overflow:trigger-noip-resume	On PMC overflow, resume and emit TRIG
on-event:trigger		On PMC event, TRIG with IP attribution
on-event:trigger-pause		On PMC event, pause and emit TRIG with ip
on-event:trigger-resume		On PMC event, resume and emit TRIG with ip
on-event:trigger-noip		On PMC event, emit TRIG packet
on-event:trigger-noip-pause	On PMC event, pause and emit TRIG
on-event:trigger-noip-resume	On PMC event, resume and emit TRIG
on-match:trigger-noip		On DR match, emit TRIG packet
on-match:trigger		On DR match, TRIG with IP attribution
on-match:trigger-pause		On DR match, pause and emit TRIG with ip
on-match:trigger-resume		On DR match, resume and emit TRIG with ip
on-match:trigger-noip-pause	On DR match, pause and emit TRIG
on-match:trigger-noip-resume	On DR match, resume and emit TRIG

where	PMC = Performance monitor Counter
	DR = Debug Register (Breakpoint)

Relevant capabilities in the intel_pt/caps directory are:

	trigger_tracing		must be 1
	num_trigger_msrs	must not be 0
	trigger_attribution	allows trigger-ip
	dr_match		allows on-match
	pause_resume		allows pause and resume

For 2 triggers, separate by "::" e.g.

on-event:trigger::on-overflow:trigger-pause

Refer to the Intel PT chapter of the Intel Architectures Software
Developer’s Manual Volume 3, for details of Trigger Tracing.

Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>