perf/x86/intel: Update idxmsk64 of 'unconstrained' structure with new parsed counter bitmap
The idxmsk64 field of 'unconstrained' indicates all supported GP
counters which can be used. For hybrid arch with multiple PMUs, only
the index mask of PMU on booting CPU is correctly initialized, the index
mask of all other PMUs on non-booting CPUs needs to be updated after
reading the counter bitmap.

Signed-off-by: Dapeng Mi <dapeng1.mi@linux.intel.com>