drm/i915/xe2lpd: Confirm if it is a tc port before checking for mode
Explicitly check if the port is a tc port before accessing the members
of intel_tc_port. If it is a non tc port, then the tc pointer could be
NULL causing kernel oops.

This check was present previously but disappeared when there was a
refactoring done in this commit -
(drm/i915/xe2lpd: move legacy code out of the main _max_lane_count() func)

The redundant check inside intel_tc_port_get_max_lane_count() is
removed.

Signed-off-by: Balasubramani Vivekanandan <balasubramani.vivekanandan@intel.com>