KVM: VMX: Update HFI table when vCPU migrates
When the vCPU migrates to a different pCPU, the virtual hfi data
corresponding to the vCPU's hfi index should be updated to the new
pCPU's data.

We don't need to re-register HFI notifier because currently ITD/HFI
virtualization is only supported for client platforms (with only one
HFI instance).

In this case, make the request to update the virtual hfi table.

Signed-off-by: Zhuocheng Ding <zhuocheng.ding@intel.com>
Co-developed-by: Zhao Liu <zhao1.liu@intel.com>
Signed-off-by: Zhao Liu <zhao1.liu@intel.com>
---
Changes since v3:
 * None.

Changes since v2:
 * Rename the title from "KVM: x86: *" to "KVM: VMX: *"..

Changes since v1:
 * Drop the logic to update notification.
 * Drop the logic to update hfi data for single vCPU because in the
   client case, the CPU is not too much, and the increased overhead is
   acceptable compared to adding more complex logic.
 * Rewrite a simpler commit message.