selftests/tdx-compliance: Add CR Compliance Tests
The Control Register (CR) is a set of registers within the processor
that contain various control bits used to manage its operations.
However, if there is a vulnerability in the reading and writing of
CR in the Intel TD environment, it may lead to security issues.

CR-compliance is to verify the behavior of the TDX module, preventing the
potential malicious happening. Add functions get_cr0 and get_cr4 saving
values of CR0 and CR4 with exception handling. Once exceptions are
triggered by tests, control flow may be returned to the instruction after
the fault instruction, and the fixup function of the trap with type
EX_TYPE_FAULT can handle it.

The tests are from chapter 11.6 (CR Handling), in the Intel® Trust
Domain Extensions (Intel® TDX) Module Base Architecture Specification.

Usage:
  insmod ./tdx-compliance.so
  echo cr > /sys/kernel/debug/tdx/tdx-tests
  cat /sys/kernel/debug/tdx/tdx-tests

Co-developed-by: Dongcheng Yan <dongcheng.yan@intel.com>
Signed-off-by: Dongcheng Yan <dongcheng.yan@intel.com>
Signed-off-by: Yi Sun <yi.sun@intel.com>