perf/x86/uncore: Use u64 to replace unsigned for the uncore offsets array
Current perf uses unsigned type for the uncore offsets array, but u64
type for the base address. The address of the later units can be
calculated by base address + offset. Perf assumes that the offset is
always positive. But it's not the case for some units on GNR. The later
units may have a lower address than the base address. Use u64 to replace
unsigned for the uncore offsets array to correct the above case.

There is no functional change.

Signed-off-by: Kan Liang <kan.liang@linux.intel.com>