KVM: TDX: add seamcall TDH_SERVTD_PREBIND
The TDH_SERVTD_PREBIND seamcall pre-binds a user TD to a service TD. A
service TD is a special TD launched in the same way as launching a regular
user TD, and its job is to provide services to the user TDs that it has
been bound to, e.g. assist in the process of live migrating the user TD.

Service TD is part of the user TD's TCB, so it needs to be included in
the user TD's attestion process. With pre-binding, a hash value of the
service TD's TDINFO_STRUCT is provided.

The main purpose of pre-binding is to facilitate the build of a transient
service TD. For example, a user TD may need services from a service TD
later in its lifecycle (e.g. services to assist live migration when needed
later). In this case, the service TD doesn't have to be launched before
launching the user TD, and it can be launched on demand when the service
is needed, and after the service is done, the service TD can be
terminated. For such usages, do prebinding with a hash of the service TD
to allow later binding with the service TD of the same hash.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>