KVM: TDX: add seamcall TDH_SERVTD_BIND
The TDH_SERVTD_BIND seamcall binds a user TD to a service TD. After the
seamcall is made successfully, the service TD can provide services to the
user TD (e.g. read and write control structure fields of a user TD using
tdcalls). Once binding is done successfully, a 64-bit binding handle and
256-bit UUID of the user TD are returned. These are needed by the service
TD when later acceessing to the user TD's control structure fields.

The TDH_SERVTD_BIND can be used in 3 cases:
- initial binding: No pre-binding has been done on the user TD. This
  requires TDH_SERVTD_BIND to be called before TDH_MR_FINALIZE.
- later initial binding: Pre-binding has been done. The binding seamcall
  can be made at any time. The service TD to bind should match to the one
  that was pre-bound (i.e. hash of TDINFO_STRUCT, attr should match).
- rebinding: binding has been done, but may need to rebind to a new
  serice TD instance, e.g. after live migration to a new physical machine.

The TDH_SERVTD_BIND seamcall requires that the service TD has been running
at the time the call is made, as the TDX module needs to find the related
info of the running service TD's to calculate a hash to
1) include into the user TD's attestaion for the initial binding case; or
2) check the if the hash matches to the one included via pre-binding for
the late initial binding case.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>