KVM: TDX: handle TDG_VP_VMCALL_SERVICE.QUERY
As defined in the GHCI spec, the QUERY service is one of the services
supported by TDG_VP_VMCALL_SERVICE. Currently only 1 command, query, is
supported, which allows the service TD to query if a service (identified
by a 128-bit service GUID) handling is supported by KVM. The 128-bit GUIDs
for various services defined in the spec are translated into enum
tdvmcall_service_id type for easy referencing.

If the queried service is supported, TDVMCALL_SERVICE_QUERY_S_SUPPORTED
is returned in the status field. Otherwise, the status field is set to
TDVMCALL_SERVICE_QUERY_S_UNSUPPORTED.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>