KVM: TDX: handle TDG_VP_VMCALL_SERVICE.MigTD and its WaitForRequest cmd
The MigTD service defined in the GHCI spec is one of the services
supported by TDG_VP_VMCALL_SERVICE. It supports a bunch of commands, which
are used by MigTD (a service TD to assisit the TD live migration setup) to
request KVM to handle.

WaitForRequest is a command used by MigTD to check from KVM if there is an
operation needs to perform, e.g. start migration. Once the operation
"start migration" is received from userspace (via the
KVM_TDX_SET_MIGRATION_INFO cmd), KVM walks through the MigTD's array of
binding slots (added by user TDs) to check if there is a slot whose
binding state is TDX_BINDING_SLOT_STATE_PREMIG_WAIT, which indicates the
slot holds all the info for MigTD and waits for pre-migration setup to
start. If such a slot is detected, the slot's state will be transit to
TDX_BINDING_SLOT_STATE_PREMIG_PROGRESS, indicating that it is being
processed by MigTD, and MigTD gets the associated data below:
- migtd_basic_info: the basic migration information, e.g. the user TD's
128-bit uuid, binding handle.
- migtd_socket_info: information related to the host socket that MigTD
will connect to, e.g. cid and vsock port.
- migtd_policy_info: information related to migration policy, e.g. if
the destination machine is in a location allowed for the TD to be
migrated. This data could be cloud vendor specific, and currently passed
with a magic number for testing purpose.

If no such a slot is detected, the MigTD's vCPU will be put in halt state
and scheduled out.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>