thermal: intel: hfi: Add a sysfs interface for enablement state
Determining whether HFI is enabled can be inferred by looking at the
MSR_IA32_HW_FEEDBACK_CONFIG register, inspecting the scheduler behavior,
or listening for thermal netlink messages. All these methods are
cumbersome and unreliable.

Create sysfs interface that unequivocally states whether the HFI is enabled
and in use.

The scope of the HFI is per package. Creating a per-package unnecessarily
complex. Instead, insert the new file under the node of individual CPUs.
CPUs belonging to the same package will show the same enablement status.

When HFI is enabled, the output of the new file is as follows:

     user@host:~$: cat /sys/devices/system/cpu/cpuX/intel_hfi
     1

and 0 otherwise.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>