thermal: intel: hfi: Add command-line options
Add command-line options to control the enablement the HFI driver. Provide
an option to forcefully enable the driver on processors knonw to have
broken HFI. Likewise, provide an option to disable the driver.

The default behavior is to enable the driver if hardware supports HFI and
is not know to be broken.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>