perf/x86/intel: Add cnt_bitmapl to represent 64 bits available counter bitmap
In previous implementation, we use a unsigned long type array to
represent the available counter bitmap, It needs extra conversion from
the array to u64 when we want to get the GP or fixed counter bitmap. So
we add a new u64 type cnt_bitmapl and put it into a union structure with
the original cnt_bitmap array. So we can read the counter bitmap
directly and avoid unnecessary conversion.

Signed-off-by: Dapeng Mi <dapeng1.mi@linux.intel.com>
Reported-by: kernel test robot <lkp@intel.com>