KVM: x86: Advertise AMX-SPARSE CPUID to user space
AMX-SPARSE is a new set of instructions in the latest Intel platform
Diamond Rapids, for AMX Sparse compute acceleration.

The bit definition:
CPUID.(EAX=7,ECX=1):EDX[bit 20]

AMX-SPARSE is on a KVM-only subleaf. Plus an x86_FEATURE definition
for this feature bit to direct it to the KVM entry.

Advertise AMX-SPARSE to KVM userspace. This is safe because there
are no new VMX controls or additional host enabling required for
guests to use this feature.

Signed-off-by: Tao Su <tao1.su@linux.intel.com>