KVM: x86: Advertise AVX512-VNNI-INT8 CPUID to user space
AVX512-VNNI-INT8 is a new set of instructions in the latest Intel
platform Diamond Rapids, for expanding VNNI support for all signed
and unsigned combinations.

The bit definition:
CPUID.(EAX=7,ECX=1):EDX[bit 2]

AVX512-VNNI-INT8 is on a KVM-only subleaf. Plus an x86_FEATURE
definition for this feature bit to direct it to the KVM entry.

Advertise AVX512-VNNI-INT8 to KVM userspace. This is safe because
there are no new VMX controls or additional host enabling required
for guests to use this feature.

Signed-off-by: Tao Su <tao1.su@linux.intel.com>