KVM: VMX: Introduce hfi table index for vCPU
The HFI table contains a table header and many table entries. Each table
entry is identified by an HFI table index, and each CPU corresponds to
one of the HFI table indexes [1].

Add hfi_table_idx in vcpu_vmx, and this will be used to build virtual
HFI table.

This HFI index is initialized to 0, but in the following patch the VMM
can be allowed to configure this index with a custom value (CPUID.0x06.
edx[bits 16-31]).

[1]: SDM, vol. 3B, section 15.6.1 Hardware Feedback Interface Table
     Structure

Signed-off-by: Zhao Liu <zhao1.liu@intel.com>
---
Changes since v3:
 * None.

Changes since v2:
 * Split hfi_table_idx support from virtual HFI table support patch to
   simplify the logic of the latter.