KVM: TDX: read the max number of servtds supported per user TD
During TDX module setup, read from the TDX module the max number of
service TDs supported for a user TD to be bound to. The info will be
referenced when KVM receives binding request from userspace, e.g.
binding will be performed only when the number of service TDs that
have been bound to ther user TD is less than the max number.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>