x86/microcode/intel: Handle legacy update when deferred commit is enabled
With introduction of rollback and deferred commit, microcode SVN
can be committed at a later point after WRMSR to 0x79 using
manual-commit or can be committed immediately as part of WRMSR 0x79
using auto-commit.

Extend legacy update via reload sysfs to use auto-commit if the
architecture permits.

In addition, this commit also refactors the code a bit to add
support for rollback and deferred commit in the upcoming patches.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>
Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>