x86/microcode: Add support for ucode late load without committing
Deferred commit allows user to test a microcode before committing it
into production. If the user experiences instability, they can
simply perform a rollback to the previously installed microcode.

In order to permit this there are a few constraints that OS needs
to check before permitting an update.

1. Current CPU should have rollback support.
2. New microcode must have the meta-data required to support rollback.
3. The meta-data must indicate current installed version as supported
   rollback candidate.
4. Adhere to SVN commit rules (please refer to spec). Only exception
   being, prevent loading if new MCU MIN_SVN > current CPU SVN as
   one cannot rollback to the previous revision.

This commit adds support for loading microcode after checking the above
constraints. In the next patch, we will add support for committing the
microcode.

This commit also sets up a new directory
`/lib/firmware/intel-ucode/stage` to load the deferred commit
microcode. This is done to avoid replacing with production microcode
in the standard filesystem path with something that is yet to be
promoted as production microcode.

Signed-off-by: Ashok Raj <ashok.raj@intel.com>
Signed-off-by: Vijay Dhanraj <vijay.dhanraj@intel.com>