platform/x86/intel/pmc: Split pmc_core_ssram_get_pmc into separate functions.
Each PMC is represented by an SSRAM device, however only the first is
discoverable as a PCI device to the OS. The BARs to the remaining PMCs are
described in memory of the first PMC, making the discovery process
slightly different. Create two separate functions for finding the primary
and secondary PMCs. Also changes the return type from void to allow
returning an error when failing to find the primary PMC.

Signed-off-by: David E. Box <david.e.box@linux.intel.com>