platform/x86/intel/sdsi_spdm: Add support for GET_MEASUREMENTS
Add support for doing GET_MEASUREMENTS. This allows retrieval of firmware
measurements (e.g. hash of ROM) after the device has been pre-authenticated
through attestation. Per the spec, measurement requests can happen multiple
times.  If signing is supported, a transcript of all of the measurement
requests will be taken and an incremental hash computed. When a signing
request is made, finalize the hash and use it to verify the signature
returned by the device (Responder). If signing is supported, assume that a
signed request can be made at any time and start calculating the hash. Per
the spec, end the transcript after,

	- a signed measurement is requested,
	- a message other than GET_MEASUREMENTS is sent,
	- an error occurs.

The leaf key obtained during spdm_authorize will be used to verify the
signature. Use a callback to send each measurement back to the caller after
is has be obtained.

Signed-off-by: David E. Box <david.e.box@linux.intel.com>