platform/x86: ISST : Support partiion id and compute die mask
A partioned system can have more than one TPMI instance per package. In
that case store SST instance for each partition for the same package as
part of isst_common.sst_inst[]. The size of isst_common.sst_inst[]
supports two partitions, which is the max number of possible partitions.

Also check the mask of compute dies in the partition to get proper
instance. This will support the case where compute die ID of a CPU is
unique for the package even in the multiple partitioned system.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>