thermal: intel: hfi: Add a PM notifier block for hibernation
The kernel gives the HFI hardware a memory region that it uses to provide
updates the HFI table. The kernel allocates this memory region during CPU
hotplug.

When resuming from suspend-to-disk, the restore kernel reprograms the HFI
hardware with a memory location that is potentially different than that
set by the image kernel. Restore the original location of the HFI memory
when restoring the system to its pre-hibernation state.

Disable HFI when transitioning in and out of hibernation and suspend-to-
RAM. Do this only for the package of the boot CPU. The HFI instances of
of other packages are handled in the CPU hotplug callbacks.

Cc: Chen Yu <yu.c.chen@intel.com>
Cc: Len Brown <len.brown@intel.com>
Cc: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Cc: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
Cc: Zhang Rui <rui.zhang@intel.com>
Cc: Zhao Liu <zhao1.liu@linux.intel.com>
Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>