Exposing apic functions that are not resolving in SVOS OOT builds
We noticed in svfs that we were getting modpost errors for some apic
functions when we built it against 6.6. These functions do get used
in svfs:

* apic_eoi
* apic_read
* apic_write
* apic_icr_read
* apic_icr_write
* apic_wait_icr_idle

We create svoskern_* versions of this functions in our kernel,
specifically in svos.c where they are protected by the CONFIG_SVOS
build cofig option. They are exported as symbols we can use in
svfs. Svfs has a matching modification to call the new svoskern_*
versions.

There is also read_apic_id is not a direct static_call, but it uses
apic_read, and is invoked as an inline function in apic.h in svfs.
So we had to guard that call to use the svoskern version as well.